requirements.txt
# Core
pydantic==2.*
apache-airflow

# Testing
pytest
pytest-cov
bump2version

# Docs
requests
pyyaml
sphinx
myst-parser
sphinxcontrib-mermaid
sphinx-rtd-theme
apache-airflow-providers-docker
